+++ tests/test_main_addheader.py
@@ -1061,3 +1061,87 @@ def test_addheader_skip_existing(fake_repository, stringio, mock_date_today):
     assert result == 0
     assert (fake_repository / "foo.py").read_text() == expected_foo
     assert (fake_repository / "bar.py").read_text() == expected_bar
+
+
+def test_addheader_recursive(fake_repository, stringio, mock_date_today):
+    """Add a header to a directory recursively."""
+    (fake_repository / "src/one/two").mkdir(parents=True)
+    (fake_repository / "src/one/two/foo.py").write_text(
+        cleandoc(
+            """
+            # spdx-License-Identifier: GPL-3.0-or-later
+            """
+        ).replace("spdx", "SPDX")
+    )
+    (fake_repository / "src/hello.py").touch()
+    (fake_repository / "src/one/world.py").touch()
+    (fake_repository / "bar").mkdir(parents=True)
+    (fake_repository / "bar/bar.py").touch()
+
+    result = main(
+        [
+            "addheader",
+            "--copyright",
+            "Joe Somebody",
+            "--recursive",
+            "src/",
+        ],
+        out=stringio,
+    )
+
+    for path in (fake_repository / "src").glob("src/**"):
+        content = path.read_text()
+        assert (
+            "spdx-FileCopyrightText: 2018 Joe Somebody".replace("spdx", "SPDX")
+            in content
+        )
+
+    assert "Joe Somebody" not in (fake_repository / "bar/bar.py").read_text()
+    assert result == 0
+
+
+def test_addheader_recursive_on_file(
+    fake_repository, stringio, mock_date_today
+):
+    """Don't expect errors when addheader is run 'recursively' on a file."""
+    result = main(
+        [
+            "addheader",
+            "--copyright",
+            "Joe Somebody",
+            "--recursive",
+            "src/source_code.py",
+        ],
+        out=stringio,
+    )
+
+    assert (
+        "Joe Somebody" in (fake_repository / "src/source_code.py").read_text()
+    )
+    assert result == 0
+
+
+def test_addheader_recursive_contains_unrecognised(
+    fake_repository, stringio, mock_date_today
+):
+    """Expect error and no edited files if at least one file has not been
+    recognised."""
+    (fake_repository / "baz").mkdir(parents=True)
+    (fake_repository / "baz/foo.py").write_text("foo")
+    (fake_repository / "baz/bar.unknown").write_text("bar")
+    (fake_repository / "baz/baz.sh").write_text("baz")
+
+    with pytest.raises(SystemExit):
+        main(
+            [
+                "addheader",
+                "--license",
+                "Apache-2.0",
+                "--copyright",
+                "Jane Doe",
+                "--recursive",
+                "baz/",
+            ]
+        )
+
+    assert "Jane Doe" not in (fake_repository / "baz/foo.py").read_text()